+++ paddle_billing/Notifications/Requests/__init__.py
@@ -1,0 +1,1 @@
+from paddle_billing.Notifications.Requests.Request import Request